test: test

+++ Backend/alcol/log-service/src/main/java/com/alcol/logservice/controller/LogController.java
@@ -92,7 +92,7 @@ public class LogController
         log.info("LogController 분기 1");
 
         String battleMode = (String)map.get("battleMode");
-        Long probNum = Long.valueOf((int)map.get("probNum"));
+        Long probNum = (Long)map.get("probNum");
 
         log.info("LogController 분기 2");
